# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NUTTER-XMD WhatsApp Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface PairRequestBody {
  /** Phone number in international format (e.g. +254712345678) */
  phoneNumber: string;
}

export interface PairRequestResponse {
  /** 8-character pair code to enter in WhatsApp */
  pairCode: string;
  phoneNumber: string;
}

export interface PairQrResponse {
  /** QR code as base64 data URL (data:image/png;base64,...) */
  qr: string;
  expiresAt?: string;
}

export type PairStatusResponseStatus =
  (typeof PairStatusResponseStatus)[keyof typeof PairStatusResponseStatus];

export const PairStatusResponseStatus = {
  idle: "idle",
  connecting: "connecting",
  qr_ready: "qr_ready",
  pair_code_ready: "pair_code_ready",
  connected: "connected",
  disconnected: "disconnected",
} as const;

export interface PairStatusResponse {
  status: PairStatusResponseStatus;
  phoneNumber?: string | null;
}

export interface PairSessionResponse {
  /** Base64-encoded session credentials to paste into Heroku config */
  sessionId: string;
  phoneNumber: string;
}

export interface ForkVerifyResponse {
  forked: boolean;
  username: string;
  forkUrl?: string | null;
  deployUrl?: string | null;
}

export interface Fork {
  id: number;
  login: string;
  avatarUrl: string;
  profileUrl: string;
  forkUrl: string;
  createdAt: string;
}

export interface ForksResponse {
  forks: Fork[];
  total: number;
}

export type VerifyForkParams = {
  username: string;
};