# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NUTTER-XMD WhatsApp Bot API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Accepts a phone number, initializes a Baileys session, returns a pair code
 * @summary Request a pairing code
 */
export const PairRequestBody = zod.object({
  phoneNumber: zod
    .string()
    .describe("Phone number in international format (e.g. +254712345678)"),
});

export const PairRequestResponse = zod.object({
  pairCode: zod.string().describe("8-character pair code to enter in WhatsApp"),
  phoneNumber: zod.string(),
});

/**
 * Returns the current QR code as a base64 data URL
 * @summary Get QR code for pairing
 */
export const GetPairQrResponse = zod.object({
  qr: zod
    .string()
    .describe("QR code as base64 data URL (data:image\/png;base64,...)"),
  expiresAt: zod.coerce.date().optional(),
});

/**
 * Returns the current connection/pairing state
 * @summary Get pairing status
 */
export const GetPairStatusResponse = zod.object({
  status: zod.enum([
    "idle",
    "connecting",
    "qr_ready",
    "pair_code_ready",
    "connected",
    "disconnected",
  ]),
  phoneNumber: zod.string().nullish(),
});

/**
 * Returns the base64-encoded session ID once WhatsApp is connected
 * @summary Get session ID after successful pairing
 */
export const GetPairSessionResponse = zod.object({
  sessionId: zod
    .string()
    .describe("Base64-encoded session credentials to paste into Heroku config"),
  phoneNumber: zod.string(),
});

/**
 * Clears the current pairing session to start fresh
 * @summary Reset pairing session
 */
export const ResetPairingResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns list of GitHub users who forked the NUTTER-XMD repo
 * @summary Get all repo forks
 */
export const GetAdminForksResponse = zod.object({
  forks: zod.array(
    zod.object({
      id: zod.number(),
      login: zod.string(),
      avatarUrl: zod.string(),
      profileUrl: zod.string(),
      forkUrl: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
  total: zod.number(),
});

/**
 * Checks if a given GitHub username has forked the NUTTER-XMD repo
 * @summary Verify GitHub fork
 */
export const VerifyForkQueryParams = zod.object({
  username: zod.coerce.string(),
});

export const VerifyForkResponse = zod.object({
  forked: zod.boolean(),
  username: zod.string(),
  forkUrl: zod.string().nullish(),
  deployUrl: zod.string().nullish(),
});